// src/app/reporte-de-ventas/components/mensuales/types.ts
export interface DatoGraficoMensual {
  mes: string;
  A: number;
  X: number;
  AX: number;
  cantidadA: number;
  cantidadX: number;
  cantidadAX: number;
  varA: number;
  varX: number;
  varAX: number;
  varCantA: number;
  varCantX: number;
  varCantAX: number;
  tieneVariacion: boolean;
  ghostValue: number;
}